OurSpace Login HTML and PHP files
I have added the html file for the login page.
Everything should work, but the design still
needs to  be worked on.

The php file has also been uploaded. It connects
our local database to our login. The user is able
to succussfully log in and go to the home page.
There still needs work to be done like password
encyryption and changing to aws MySQL