Login HTML and PHP Attempt #2
Had to retry and upload the files, since
last time the files did not upload correctly

The html file works fine, however the design
still needs to be worked on. The php file
is also connected to the local database and works
perfectly. However, we need to work on password
encryption and making our database on AWS.