Working Database with JS files, Small Form and CSS fixes
I have made and connected the database to our user
login/signup inputs. Once the user inputs the data
in the signup page they will be redirected to the
login page and their information will be stored in
our database. The user will then be able to
successfully be able to login and will be redirected
to the login page.

Fixes:
- Specify that username should be an email
- Maybe add an extra input for a username so our users
are not stuck with their emails as their display name