Completed User Profile and Fixes:
I have completed the user profile. The user is able to
successfully see their username on screen and their bio
if they have one. The user is also able to see their
profile color. All the buttons are wokring besides two
(post and add friend). When the user clicks edit, they
are able to customize their profile and be able to
save their changes.

I also fixed the homepage so that some of the buttons work.
To add on, I made a file (one of the peer review suggestions)
to configure firebase. That way I will not have to
keep on copying and pasting a whole chunck of code.

Issues:
1. Need to add user posts
2. Add the remaining buttons