Added Profile Images
I added the ability for the users to add their own
profile images, instead of colors. The user can choose any
image from their computer and it will be uploaded to
our storage. I also added the ability for the user to
be able to crop their images to display what they want.

Added
- Profile images
- crop
- deleted color